Update package import path
Go modules need to have a major version suffix in the import path.
This is enforced for repositories that contain a go.mod file.

https://go.dev/ref/mod#major-version-suffixes

Go-ovirt introduced a go.mod file in v4.3.5, which means the Go tools
still download v4.3.4 when you add the dependency.

This commit adds the /v4 suffix to go.mod and all examples.

Signed-off-by: Vic Demuzere <vic@demuzere.be>